- LocationManagerSensorReader  get locations from GPS and computes time interval between consecutive updates - GpsVelocityCalculator compute distance between two locations (Location.distanceBetween()) and calculate average speed using known time interval between location readings.

+++ app/src/main/java/alex_shutov/com/ledlights/sensor/Reading.java
@@ -10,17 +10,17 @@ import android.hardware.SensorEvent;
  * Readings wrapper, substract drift from values
  */
 public class Reading {
-    public float values[];
+    public double values[];
     public long timeInterval;
     // rarely used
     public long timestamp;
 
     public Reading() {
-        values = new float[3];
+        values = new double[3];
     }
 
     public Reading(SensorEvent event, long timeInterval) {
-        values = new float[3];
+        values = new double[3];
         values[0] = event.values[0];
         values[1] = event.values[1];
         values[2] = event.values[2];
